TIDIJ-62 Created dedicated class file for UDPServer
